Refactor AppDocumentation class
