+++ lib/apiculture/markdown_segment.rb
@@ -3,4 +3,8 @@ class Apiculture::MarkdownSegment < Struct.new(:string)
   def to_markdown
     string.to_s
   end
+
+  def to_markdown_slice(_mountpoint)
+    TaggedMarkdown.new(self, 'apiculture-verbatim')
+  end
 end
